Break out evidence config.
